refactor: modularize helper functions into own file
